Moved eval files to angel_system, added ptg_eval to poetry
