+++ pyproject.toml
@@ -19,6 +19,9 @@ transitions = "^0.8.11"
 timm = "^0.5.4"
 einops = "^0.4.1"
 pynput = "^1.7.6"
+scriptconfig = "^0.6.4"
+pandas = "^1.4.3"
+pyarrow = "^9.0.0"
 
 [tool.poetry.dev-dependencies]
 ipython = "^8.0.1"
@@ -26,6 +29,9 @@ ipython = "^8.0.1"
 [build-system]
 requires = ["poetry-core==1.0.8"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.scripts]
+ptg_eval = "angel_system.eval.ptg_evaluate:main"
 
 [tool.poetry.plugins."smqtk_plugins"]
 # Activity detector
